Refactor NotificationScreen
